requirements.txt
# ------------------------------------------------------------------
# Core deep-learning stack
# ------------------------------------------------------------------
torch==2.2.2
torchvision==0.17.2

# ------------------------------------------------------------------
# Diffusion pipeline (Stable Diffusion 1.5)
# ------------------------------------------------------------------
diffusers==0.27.2
transformers==4.40.1
accelerate==0.29.3
safetensors==0.4.3

# ------------------------------------------------------------------
# Image processing
# ------------------------------------------------------------------
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0

# ------------------------------------------------------------------
# Web API (FastAPI server)
# ------------------------------------------------------------------
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# ------------------------------------------------------------------
# HuggingFace model download helper
# ------------------------------------------------------------------
huggingface-hub==0.23.0
